+++ app/src/main/java/com/greenfox/kalendaryo/http/backend/MockBackendApi.java
@@ -3,9 +3,9 @@ package com.greenfox.kalendaryo.http.backend;
 import com.greenfox.kalendaryo.models.GoogleAuth;
 import com.greenfox.kalendaryo.models.Kalendar;
 import com.greenfox.kalendaryo.models.KalUser;
-import com.greenfox.kalendaryo.models.GetKalendarListResponse;
-import com.greenfox.kalendaryo.models.GetKalendarResponse;
-import com.greenfox.kalendaryo.models.PostKalendarResponse;
+import com.greenfox.kalendaryo.models.responses.GetKalendarListResponse;
+import com.greenfox.kalendaryo.models.responses.GetKalendarResponse;
+import com.greenfox.kalendaryo.models.responses.PostKalendarResponse;
 
 import java.util.ArrayList;
 import java.util.Arrays;
