+++ app/src/main/java/com/greenfox/kalendaryo/models/responses/GetKalendarResponse.java
@@ -1,4 +1,4 @@
-package com.greenfox.kalendaryo.models;
+package com.greenfox.kalendaryo.models.responses;
 
 import java.util.List;
