+++ app/src/main/java/com/greenfox/kalendaryo/models/responses/GoogleCalendarsResponse.java
@@ -1,4 +1,6 @@
-package com.greenfox.kalendaryo.models;
+package com.greenfox.kalendaryo.models.responses;
+
+import com.greenfox.kalendaryo.models.GoogleCalendar;
 
 import java.util.List;
